Tweak line plotting instructions in topcat tutorial
Add instruction to turn off the point plotting at the same time as
turning on the array plotting. This doesn't matter in this particular
example, but if there are some scalar columns in the same table as the
array columns, not doing it can cause the plot auto-scaling to display
the wrong part of the plane (which typically scrunches the spectral
plots into one corner).